Verifing if users schema exists

+++ test/test_helper.js
@@ -5,8 +5,10 @@ mongoose.Promise = global.Promise
 before((done) => {
     //development mode
     //DATABASE_API = mongodb://localhost
-    mongoose.connect(`${process.env.DATABASE_API}/users_test`, { useNewUrlParser: true, 
-        useUnifiedTopology: true, useFindAndModify: false})
+    mongoose.connect(`${process.env.DATABASE_API}/users_test`, {
+        useNewUrlParser: true,
+        useUnifiedTopology: true, useFindAndModify: false
+    })
     mongoose.connection
         .once('open', () => { done() })
         .on('error', (error) => {
@@ -15,11 +17,15 @@ before((done) => {
 })
 
 beforeEach((done) => {
-    const { users} = mongoose.connection.collections
-    users.drop()
-    .then(() => done())
-    .catch(error=>{
-        console.log(error)
-        done(error)
-    }) 
+    if (mongoose.connection.collections.users) {
+        const { users } = mongoose.connection.collections
+        users.drop()
+            .then(() => done())
+            .catch(error => {
+                console.log(error)
+                done(error)
+            })
+    } else {
+        done()
+    }
 })
